Fix path issues in local ad, fix issues in AdSerializer
Several issues were fixed in the LocalAd class. First, the local data is now organized per user to keep the system functioning if two users sign in the app from the same phone. There was also an issue with using pairs inside a unit test, this has been resolved by not using pairs at all. Fixed an issue where we retrieved the photo refs from the deserialized ad when they were in fact, only contained in the HashMap we retrieved from the file.